Restructure Infinite Pages. Self contain for deprecation
