+++ frontends/mit-open/src/pages/InfinitePages/common/infinite-pages-urls.ts
