+++ frontends/mit-open/src/pages/InfinitePages/components/FieldAvatar.test.tsx
