+++ frontends/mit-open/src/pages/InfinitePages/components/FieldAvatar.tsx
