+++ frontends/mit-open/src/pages/InfinitePages/page-components/EditFieldAppearanceForm.tsx
@@ -4,10 +4,10 @@ import { Formik, Form, Field } from "formik"
 import { Button } from "ol-design"
 import * as Yup from "yup"
 
-import { FieldChannelAppearanceForm, useMutateField } from "../../api/fields"
-import type { FieldChannel } from "../../api/fields/interfaces"
-import FieldAvatar from "./components/FieldAvatar"
-import { makeFieldViewPath } from "../urls"
+import { FieldChannelAppearanceForm, useMutateField } from "../../../api/fields"
+import type { FieldChannel } from "../../../api/fields/interfaces"
+import FieldAvatar from "../components/FieldAvatar"
+import { makeFieldViewPath } from "../common/infinite-pages-urls"
 
 type FormProps = {
   field: FieldChannel
