+++ frontends/mit-open/src/pages/InfinitePages/pages/EditFieldPage.test.tsx
@@ -1,7 +1,7 @@
-import { renderTestApp, screen, setMockResponse } from "../../test-utils"
-import * as factory from "../../api/fields/test-utils/factories"
-import { FieldChannel, urls } from "../../api/fields"
-import { urls as lrUrls } from "../../api/learning-resources"
+import { renderTestApp, screen, setMockResponse } from "../../../test-utils"
+import * as factory from "../../../api/fields/test-utils/factories"
+import { FieldChannel, urls } from "../../../api/fields"
+import { urls as lrUrls } from "../../../api/learning-resources"
 
 describe("EditFieldPage", () => {
   const setup = (overrides?: Partial<FieldChannel>) => {
