+++ frontends/mit-open/src/pages/InfinitePages/pages/FieldPage.tsx
@@ -15,11 +15,11 @@ import { Link } from "react-router-dom"
 import FieldPageSkeleton from "./FieldPageSkeleton"
 import ArrowForward from "@mui/icons-material/ArrowForward"
 import ArrowBack from "@mui/icons-material/ArrowBack"
-import { useFieldDetails } from "../../api/fields"
-import { useUserListItems } from "../../api/learning-resources"
-import WidgetsList from "./WidgetsList"
-import { GridColumn, GridContainer } from "../../components/layout"
-import LearningResourceCard from "../../components/LearningResourceCard"
+import { useFieldDetails } from "../../../api/fields"
+import { useUserListItems } from "../../../api/learning-resources"
+import WidgetsList from "../components/WidgetsList"
+import { GridColumn, GridContainer } from "../../../components/layout"
+import LearningResourceCard from "page-components/LearningResourceCard/LearningResourceCard"
 import invariant from "tiny-invariant"
 
 type RouteParams = {
